Push Quandl and Batch changes
Purpose of this push is to upload the code for Quandl, scripts to form
various commodity tables and also a common commodity table. Finally,
scheduler jobs to run an hourly job to compare bitcoin transaction and
bitcoin tweet traffic. Also, a scheduler batch job to be run once a day
in order to update the commodities with daily prices.